Refactors loadCustomers
Moves the resetting of input fields into it's own method,
reset_the_input_fields.
Moves the method used to get the DOM node into it's own method, element_by_id,
making it easier to change the access method later.
Moves the retrieval and validity testing of the customers endpoint url into it's
own method, get_customers_url, since it's not part of the core logic of loading
customers.
Moves the creation of the XHR request into another new method,
request_customers, since it's also an implementation detail.

The reason to split all of this functionality out into seperate methods id so
that the main story of loading customers becomes clearer. It's now three simple
steps; clearing the form, getting the url to send the request to and sending the
request.

Fewer lines in the body of a method is normally better. Hiding implementation
details and verification in new methods with names that describe what they do,
not how, makes the flow more obvious and easier to keep in your head since there
is less noice around you.

The resetting of input fields had an actual comment saying what that block of
code did. This is a super example of something that should be it's own method.
It's not central to the story of loading customers, it's not clear on its own
and we don't care about it in that context anyway.