Refactors populateCustomerData
Extracts methods for getting the selected customer and setting the customers
data in the form.

The basic thought is the same here; get the implementation details out of the
core story. Right now you have a select box to allow for selecting a customer,
but you could just as well allow for a text field where someone could write a
customer id, and it would not affect the main story.