Refactors form filling.
Since both setting the customer data and clearing the form is essentially the
same thing we make it exactly the same thing and only use data to drive the
differance.

Looking at this from a bigger picture you should probably split out an adapter
layer that deals with both the domain objects, like customer, and the DOM.

Using that approach this operation would simply be that of sending a customer to
the adapter. The adapter is then responsible for knowing the mapping between the
domain objects and form fields. Clearing the form in this instance would just be
sending a null customer to the adapter.

The main benifit would be to split the domain logic from the presentation logic.
The customer model et.al. would be pure domain objects, encapsulating only
behaviour specific to the business domain, and the adapter would the act as a
decorator; taking a domain object and preparing it for presentation in some way.

I leave that as an exercise for the viewer for now :) Try it and I'll review it.