Refactors updateCustomerOptionsDOM
Replaces the manual HTML creation from text with real node instansiation.

The main point here is to not use string interpolation for anything that has a
real API. There is no tools for checking or validating your string in the editor
and it won't give any meaningful errors. Also it's just plain bad for
maintainability.

Using the built in methods for node creation and a few helper methods you get
the same end result but a much better stroy.