Refactors fillDhlForm
Fixes some style issues.

This method is just wrong :)

The correct way to handle this is to det up a content script, that is the script
in a chrome extention that is allowed to change the host page directly. Using
sendmessage you can then communicate between the content script and the other
parts of your extension.

So in this case the DHL form fill method would just trigger a send message with
the relevant data and the content script listens to it and updates the form on
the page.

This also gives you a nice way to isolate the parts of the extention that
depends on the underlying page from the parts that don't.

Read more here: https://developer.chrome.com/extensions/content_scripts
And look at the video at the bottom for details on how to pass messages.